+++ packages/site/src/content/AnimatedSwitcher/index.tsx
@@ -1,5 +1,6 @@
 import AnimatedSwitcherContent from "@atlantis/docs/components/AnimatedSwitcher/AnimatedSwitcher.stories.mdx";
 import MobileProps from "./AnimatedSwitcher.props.json";
+import Notes from "./AnimatedSwitcherNotes.mdx";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
@@ -38,4 +39,5 @@ export default {
       ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;