+++ packages/site/src/content/Banner/index.tsx
@@ -1,6 +1,7 @@
 import BannerContent from "@atlantis/docs/components/Banner/Banner.stories.mdx";
 import Props from "./Banner.props.json";
 import MobileProps from "./Banner.props-mobile.json";
+import Notes from "./BannerNotes.mdx";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
@@ -33,4 +34,5 @@ export default {
       ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;