+++ packages/site/src/content/LightBox/index.tsx
@@ -1,5 +1,6 @@
 import Content from "@atlantis/docs/components/LightBox/LightBox.stories.mdx";
 import Props from "./LightBox.props.json";
+import Notes from "./LightBoxNotes.mdx";
 import { ContentExport } from "../../types/content";
 import { getStorybookUrl } from "../../layout/getStorybookUrl";
 
@@ -45,4 +46,5 @@ export default {
       ),
     },
   ],
+  notes: () => <Notes />,
 } as const satisfies ContentExport;