+++ packages/site/src/layout/Layout.tsx
@@ -1,9 +1,10 @@
-import { PropsWithChildren, useEffect } from "react";
-import { Route, Switch, useLocation } from "react-router";
+import React, { PropsWithChildren, useEffect, useRef } from "react";
+import { Route, Switch, useHistory, useLocation } from "react-router";
 import { NavMenu } from "./NavMenu";
 import { AtlantisRoute, routes } from "../routes";
 import "./code-theme.css";
 import { ToggleThemeButton } from "../components/ToggleThemeButton";
+import { hooksList } from "../hooksList";
 
 /**
  * Layout for whole application. This will display the NavMenu and the content of the page.
@@ -12,17 +13,43 @@ import { ToggleThemeButton } from "../components/ToggleThemeButton";
 
 export const Layout = () => {
   const location = useLocation();
+  const scrollPane = useRef<HTMLDivElement>(null);
+  const path = new URLSearchParams(location.search).get("path");
+  const history = useHistory();
 
   useEffect(() => {
-    window.scrollTo({ top: 0 });
-  }, [location]);
+    if (scrollPane?.current) {
+      scrollPane?.current.scrollTo({ top: 0 });
+    }
+  }, [location, scrollPane?.current]);
+
+  // Redirects for the links on the hooks packages page
+  if (path && path.includes("hooks")) {
+    const pathRegex = /hooks-(.*)--docs/g.exec(path);
+    const match = hooksList.find(
+      hook => pathRegex?.[1] === hook.title.toLowerCase(),
+    );
+
+    if (match) {
+      history.push(match.to);
+    }
+  }
 
   return (
     <LayoutWrapper>
-      <NavMenu />
-      <div style={{ overflow: "auto", width: "100%", minHeight: "100%" }}>
+      <NavMenu mainContentRef={scrollPane} />
+      <div
+        style={{
+          overflow: "auto",
+          width: "100%",
+          height: "100dvh",
+          outline: "transparent",
+        }}
+        ref={scrollPane}
+        tabIndex={0}
+      >
         <Switch>
-          <>
+          <React.Fragment key={location.pathname}>
             {routes?.map((route, routeIndex) => {
               const iterateSubMenu = (childroutes: AtlantisRoute[]) => {
                 return childroutes.map((child, childIndex) => {
@@ -43,7 +70,7 @@ export const Layout = () => {
               // Top level items with children (Changelog)
               if (route.children) {
                 return (
-                  <>
+                  <React.Fragment key={route.path}>
                     <Route
                       key={routeIndex}
                       exact={route.exact ?? false}
@@ -51,7 +78,7 @@ export const Layout = () => {
                       component={route.component}
                     />
                     {iterateSubMenu(route.children)}
-                  </>
+                  </React.Fragment>
                 );
               }
 
@@ -65,7 +92,7 @@ export const Layout = () => {
                 />
               );
             })}
-          </>
+          </React.Fragment>
         </Switch>
       </div>
 
@@ -79,7 +106,7 @@ export const LayoutWrapper = ({ children }: PropsWithChildren) => {
     <div
       style={{
         display: "flex",
-        background: "var(--color-surface)",
+        background: "var(--color-surface--background)",
       }}
     >
       {children}