+++ packages/site/src/layout/NavMenu.tsx
@@ -1,133 +1,179 @@
-import { Box, Button, Typography } from "@jobber/components";
-import { Link } from "react-router-dom";
-import { Fragment, PropsWithChildren, useState } from "react";
+import { Box, Button, Content, Icon, Typography } from "@jobber/components";
+import { Link, useLocation } from "react-router-dom";
+import { Fragment, PropsWithChildren, useRef, useState } from "react";
 import { SearchBox } from "./SearchBox";
 import AnimatedPresenceDisclosure from "./AnimatedPresenceDisclosure";
+import styles from "./NavMenu.module.css";
 import { routes } from "../routes";
 import { JobberLogo } from "../assets/JobberLogo.svg";
 import { useAtlantisSite } from "../providers/AtlantisSiteProvider";
+import { VisibleWhenFocused } from "../components/VisibleWhenFocused";
+
+interface NavMenuProps {
+  readonly mainContentRef: React.RefObject<HTMLDivElement>;
+}
 
 /**
  * Left side navigation menu for the application.
  * @returns ReactNode
  */
-export const NavMenu = () => {
+export const NavMenu = ({ mainContentRef }: NavMenuProps) => {
   const [open, setOpen] = useState(false);
   const { isMinimal } = useAtlantisSite();
+  const { pathname } = useLocation();
+  const selectedRef = useRef<HTMLAnchorElement | null>(null);
 
   if (isMinimal) return null;
 
+  interface MenuItem {
+    handle: string;
+    children?: MenuItem[];
+    path?: string;
+  }
+
+  const iterateSubSubMenu = (menuItems: MenuItem[], routeIndex: number) => {
+    return menuItems.map((menuItem, menuItemIndex) => {
+      return (
+        <MenuSubItem key={`${routeIndex}-${menuItemIndex}`}>
+          <StyledSubLink
+            to={`/components/${menuItem.handle}`}
+            selectedRef={selectedRef}
+          >
+            {menuItem.handle}
+          </StyledSubLink>
+        </MenuSubItem>
+      );
+    });
+  };
+
+  const iterateSubMenu = (menuItems: MenuItem[], routeIndex: number) => {
+    return menuItems.map((menuItem, menuItemIndex) => {
+      if (menuItem.children) {
+        return (
+          <Fragment key={`${routeIndex}-${menuItemIndex}`}>
+            {sectionTitle(menuItem.handle)}
+            {iterateSubSubMenu(menuItem.children, routeIndex)}
+          </Fragment>
+        );
+      }
+
+      return (
+        <MenuSubItem key={`${routeIndex}-${menuItemIndex}`}>
+          <StyledSubLink to={menuItem.path ?? "/"} selectedRef={selectedRef}>
+            {menuItem.handle}
+          </StyledSubLink>
+        </MenuSubItem>
+      );
+    });
+  };
+
+  const skipToContent = () => {
+    mainContentRef.current?.focus();
+  };
+
   return (
-    <div
-      style={{
-        width: 220,
-        height: "100dvh",
-        backgroundColor: "var(--color-surface--background)",
-        overflow: "auto",
-      }}
-    >
-      <Box height={24} padding="base">
-        <Link to="/">
-          <JobberLogo />
-        </Link>
-      </Box>
-      <Box padding="base">
-        <Button
-          onClick={() => setOpen(true)}
-          label="Search"
-          icon="search"
-          variation="subtle"
-        />
-      </Box>
-      <SearchBox open={open} setOpen={setOpen} />
-      <MenuList>
+    <nav className={styles.navMenuContainer}>
+      <div className={styles.navMenuHeader}>
+        <VisibleWhenFocused>
+          <Button label="Skip to Content" onClick={skipToContent} />
+        </VisibleWhenFocused>
         <Box>
-          {routes?.map((route, routeIndex) => {
-            if (route.inNav === false) return null;
-
-            interface MenuItem {
-              handle: string;
-              children?: MenuItem[];
-              path?: string;
-            }
-
-            const iterateSubSubMenu = (menuItems: MenuItem[]) => {
-              return menuItems.map((menuItem, menuItemIndex) => {
-                return (
-                  <MenuItem key={`${routeIndex}-${menuItemIndex}`}>
-                    <StyledLink to={`/components/${menuItem.handle}`}>
-                      {menuItem.handle}
-                    </StyledLink>
-                  </MenuItem>
-                );
-              });
-            };
-
-            const iterateSubMenu = (menuItems: MenuItem[]) => {
-              return menuItems.map((menuItem, menuItemIndex) => {
-                if (menuItem.children) {
-                  return (
-                    <Fragment key={`${routeIndex}-${menuItemIndex}`}>
-                      {sectionTitle(menuItem.handle)}
-                      {iterateSubSubMenu(menuItem.children)}
-                    </Fragment>
-                  );
-                }
-
-                return (
-                  <MenuItem key={`${routeIndex}-${menuItemIndex}`}>
-                    <StyledLink to={menuItem.path ?? "/"}>
-                      {menuItem.handle}
-                    </StyledLink>
-                  </MenuItem>
-                );
-              });
-            };
-
-            if (route.children) {
-              return (
-                <Box key={routeIndex} padding="base">
-                  <AnimatedPresenceDisclosure
-                    to={route.path ?? "/"}
-                    title={route.handle}
-                  >
-                    {iterateSubMenu(route.children)}
-                  </AnimatedPresenceDisclosure>
-                </Box>
-              );
-            }
-
-            return (
-              <MenuItem key={routeIndex}>
-                <StyledLink to={route.path ?? "/"}>{route.handle}</StyledLink>
-              </MenuItem>
-            );
-          })}
+          {/* TODO: remove ?new when we roll out the new docs site to everyone */}
+          <Link to="/?new">
+            <JobberLogo />
+          </Link>
         </Box>
-      </MenuList>
-    </div>
+        <Box>
+          <button
+            type="button"
+            onClick={() => setOpen(true)}
+            className={styles.searchButton}
+            aria-label="Search"
+          >
+            <Icon name="search" />
+            <span className={styles.searchButtonText}>
+              <Typography
+                size={"base"}
+                textColor={"text"}
+                fontWeight={"semiBold"}
+              >
+                Search
+              </Typography>
+            </span>
+            <div className={styles.searchKeyIndicator}>/</div>
+          </button>
+        </Box>
+        <SearchBox open={open} setOpen={setOpen} />
+      </div>
+      <div className={styles.navMenu}>
+        <MenuList>
+          <Content spacing="smaller">
+            {routes?.map((route, routeIndex) => {
+              if (route.inNav === false) return null;
+
+              if (route.children) {
+                return (
+                  <Box key={routeIndex}>
+                    <AnimatedPresenceDisclosure
+                      to={route.path ?? "/"}
+                      title={route.handle}
+                      selected={pathname.startsWith(route.path ?? "/")}
+                    >
+                      {iterateSubMenu(route.children, routeIndex)}
+                    </AnimatedPresenceDisclosure>
+                  </Box>
+                );
+              }
+
+              return (
+                <MenuItem key={routeIndex}>
+                  <StyledLink
+                    to={getRoutePath(route.path) ?? "/"}
+                    selectedRef={selectedRef}
+                  >
+                    {route.handle}
+                  </StyledLink>
+                </MenuItem>
+              );
+            })}
+          </Content>
+        </MenuList>
+      </div>
+    </nav>
   );
+};
+
+const getLinkClassName = (
+  baseClasses: string,
+  isSelected: boolean,
+  selectedClass: string,
+): string => {
+  return `${baseClasses} ${isSelected ? selectedClass : ""}`.trim();
 };
 
 export const StyledLink = ({
   to,
   children,
-  style,
-}: PropsWithChildren<{ readonly to: string; readonly style?: object }>) => {
+  selectedRef,
+}: PropsWithChildren<{
+  readonly to: string;
+  readonly selectedRef: React.RefObject<HTMLAnchorElement>;
+}>) => {
+  const { pathname } = useLocation();
+  // The `to === "/?new"` check allows Home to be highlighted in the side nav
+  // TODO: remove when we roll out the new docs site & double check Home is still highlighted
+  const isSelected = pathname === to || (pathname === "/" && to === "/?new");
+  const className = getLinkClassName(
+    `${styles.navMenuItem} ${styles.navMenuLink}`,
+    isSelected,
+    styles.selected,
+  );
+
   return (
     <Link
       to={to ?? "/"}
-      style={{
-        outline: "transparent",
-        color: "var(--color-heading)",
-        fontSize: "var(--typography--fontSize-large)",
-        fontWeight: 600,
-        width: "100%",
-        textDecoration: "none",
-        userSelect: "none",
-        transition: "all var(--timing-base) ease-out",
-        ...style,
-      }}
+      className={className}
+      ref={isSelected ? selectedRef : null}
     >
       {children}
     </Link>
@@ -137,25 +183,24 @@ export const StyledLink = ({
 export const StyledSubLink = ({
   to,
   children,
-  style,
-}: PropsWithChildren<{ readonly to: string; readonly style?: object }>) => {
+  selectedRef,
+}: PropsWithChildren<{
+  readonly to: string;
+  readonly selectedRef: React.RefObject<HTMLAnchorElement>;
+}>) => {
+  const { pathname } = useLocation();
+  const isSelected = pathname === to;
+  const className = getLinkClassName(
+    `${styles.navMenuItem} ${styles.navMenuSubItem} ${styles.navMenuLink}`,
+    isSelected,
+    styles.selected,
+  );
+
   return (
     <Link
       to={to ?? "/"}
-      style={{
-        outline: "transparent",
-        color: "var(--color-heading)",
-        fontSize: "var(--typography--fontSize-base)",
-        fontWeight: 700,
-        width: "100%",
-        textDecoration: "none",
-        userSelect: "none",
-        transition: "all var(--timing-base) ease-out",
-        display: "block",
-        padding:
-          "var(--space-smaller) var(--space-smaller) var(--space-smaller) var(--space-base)",
-        ...style,
-      }}
+      className={className}
+      ref={isSelected ? selectedRef : null}
     >
       {children}
     </Link>
@@ -163,34 +208,42 @@ export const StyledSubLink = ({
 };
 
 export const MenuList = ({ children }: PropsWithChildren) => {
-  return <ul style={{ listStyleType: "none", padding: 0 }}>{children}</ul>;
+  return <ul style={{ listStyle: "none", padding: 0 }}>{children}</ul>;
 };
 
 export const MenuItem = ({ children }: PropsWithChildren) => {
   return (
-    <li
-      style={{
-        display: "flex",
-        margin: "0 var(--space-small) var(--space-smaller) var(--space-small)",
-        padding: "var(--space-small)",
-        borderRadius: "var(--radius-small)",
-        color: "var(--color-heading)",
-        alignItems: "center",
-      }}
-    >
-      {children}
+    <li style={{ listStyle: "none" }}>
+      <Typography fontWeight="bold" size="large" textColor="text">
+        {children}
+      </Typography>
     </li>
   );
 };
 
-export const changelogTitle = (
-  <Typography fontWeight="semiBold" size="large" textColor="heading">
-    Changelog
-  </Typography>
-);
+export const MenuSubItem = ({ children }: PropsWithChildren) => {
+  return (
+    <li style={{ listStyle: "none" }}>
+      <Typography fontWeight="semiBold" size="base" textColor="text">
+        {children}
+      </Typography>
+    </li>
+  );
+};
 
 export const sectionTitle = (section: string) => (
-  <Typography fontWeight="bold" size="small" textColor="textSecondary">
-    {section.toUpperCase()}
-  </Typography>
+  <div className={`${styles.navMenuItem} ${styles.navMenuSubTitle}`}>
+    <Typography fontWeight="bold" size="small" textColor="textSecondary">
+      {section.toUpperCase()}
+    </Typography>
+  </div>
 );
+
+// TODO: delete this once we roll out the new docs site to everyone
+function getRoutePath(path?: string) {
+  if (path === "/") {
+    return "/?new";
+  }
+
+  return path;
+}