+++ packages/site/src/pages/DesignPage.tsx
@@ -8,10 +8,9 @@ export const DesignPage = () => {
         header: {
           title: "Design",
           body: "The foundation of Jobber's look and feel",
-          imageURL: "../public/img-page-divider-collage.webp",
+          imageURL: "/img-page-divider-collage.webp",
         },
         body: {
-          title: "Design",
           content: designList,
         },
       }}